Fix duplicate debug info line
