+++ runepy/__init__.py
@@ -1,6 +1,7 @@
 """RunePy package."""
 
 from . import pathfinding
+from .base_app import BaseApp
 
-__all__ = ["pathfinding"]
+__all__ = ["pathfinding", "BaseApp"]
